+++ src/Ode.jl
@@ -67,7 +67,7 @@ function daedalus_ode!(du::Array, u::Array, p::Params, t::Number)
     cm_scaling = ones(p.settings) # NOTE: placeholder for future operations
     weighted_slice_sum!(p.contacts, cm_scaling, p.cm_temp)
 
-    foi = p.beta_now * p.cm_temp * community_infectious
+    foi = p.beta * p.cm_temp * community_infectious
 
     # NOTE: element-wise multiplication
     new_I = S .* foi
@@ -106,7 +106,7 @@ function daedalus_ode!(du::Array, u::Array, p::Params, t::Number)
     @. dIa = (p.sigma * E) * (1.0 - p.p_sigma) - (p.gamma_Ia * Ia)
 
     # change in hospitalised
-    @. dH = (p.eta .* Is) - ((p.gamma_H + p.omega_now) .* H)
+    @. dH = (p.eta .* Is) - ((p.gamma_H + p.omega) .* H)
 
     # change in recovered
     @. dR = (p.gamma_Ia * Ia) + (p.gamma_Is * Is) +
@@ -116,7 +116,7 @@ function daedalus_ode!(du::Array, u::Array, p::Params, t::Number)
     @. dR[:, 2] += (new_Rvax * nu_eff - new_Rwane * p.psi)
 
     # change in dead
-    @. dD = p.omega_now .* H
+    @. dD = p.omega .* H
 
     # Rt is updated by callbacks; keep its derivative zero between updates
     du[end] = 0.0